Create   Chat-App Ui.dart
This repository contains a simple Flutter project demonstrating the basic structure of a chat application UI. The app features a clean, modern design with a ListView displaying messages, including a circular avatar, user name, subtitle, and time.

The UI is built with Flutter's stateful widgets and uses Material Design principles for an intuitive and consistent user interface. The background color is set to cyan accent, with a grey app bar that hosts the title of the app, "Chat-App". This project is an excellent starting point for anyone looking to build a more advanced chat application by expanding on this simple interface...